Fix the suggested changes
