--- langlib/lang.compilertest/src/main/ballerina/src/lang.compilertest/compilertest.bal
@@ -1,6 +1,0 @@
-
-public function assertValueEqual(anydata expected, anydata actual) = external;
-public function assertError(anydata|error value) = external;
-public function assertNotError(anydata|error value) = external;
-public function assertTrue(boolean value) = external;
-public function assertFalse(boolean value) = external;
