+++ langlib/lang.test/build.gradle
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -32,4 +32,4 @@ createBalo {
     jvmTarget = 'true'
 }
 
-description = 'ballerina/lang.compilertest Implementation'
+description = 'ballerina/lang.test Implementation'
