+++ langlib/lang.test/src/main/java/org/ballerinalang/langlib/test/AssertError.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.langlib.compilertest;
+package org.ballerinalang.langlib.test;
 
 import org.ballerinalang.jvm.BallerinaErrors;
 import org.ballerinalang.jvm.scheduling.Strand;
@@ -28,10 +28,10 @@ import org.ballerinalang.natives.annotations.BallerinaFunction;
 /**
  * Native implementation of assertError(anydata|error value).
  *
- * @since 2.0
+ * @since 1.3.0
  */
 @BallerinaFunction(
-        orgName = "ballerina", packageName = "lang.compilertest", functionName = "assertError",
+        orgName = "ballerina", packageName = "lang.test", functionName = "assertError",
         args = {@Argument(name = "value", type = TypeKind.UNION)},
         isPublic = true
 )
