+++ langlib/lang.test/src/main/java/org/ballerinalang/langlib/test/AssertNotError.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2020, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,28 +16,29 @@
  * under the License.
  */
 
-package org.ballerinalang.langlib.compilertest;
+package org.ballerinalang.langlib.test;
 
 import org.ballerinalang.jvm.BallerinaErrors;
 import org.ballerinalang.jvm.scheduling.Strand;
+import org.ballerinalang.jvm.values.ErrorValue;
 import org.ballerinalang.model.types.TypeKind;
 import org.ballerinalang.natives.annotations.Argument;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 
 /**
- * Native implementation of assertTrue(boolean value).
+ * Native implementation of assertNotError(anydata|error value).
  *
- * @since 2.0
+ * @since 1.3.0
  */
 @BallerinaFunction(
-        orgName = "ballerina", packageName = "lang.compilertest", functionName = "assertTrue",
-        args = {@Argument(name = "value", type = TypeKind.BOOLEAN)},
+        orgName = "ballerina", packageName = "lang.test", functionName = "assertNotError",
+        args = {@Argument(name = "value", type = TypeKind.UNION)},
         isPublic = true
 )
-public class AssertTrue {
-    public static void assertTrue(Strand strand, boolean value) {
-        if (!value) {
-            throw BallerinaErrors.createError("Not True");
+public class AssertNotError {
+    public static void assertNotError(Strand strand, Object value) {
+        if (value instanceof ErrorValue) {
+            throw BallerinaErrors.createError("Error");
         }
     }
 }
