+++ langlib/langlib-test/build.gradle
@@ -52,7 +52,7 @@ dependencies {
     testRuntime project(':ballerina-lang:value')
     testRuntime project(':ballerina-lang:xml')
     testRuntime project(':ballerina-lang:bool')
-    testRuntime project(':ballerina-lang:compilertest')
+    testRuntime project(':ballerina-lang-test')
     testRuntime project(':ballerina-runtime')
     testRuntime project(':ballerina-lang:internal')
     testRuntime 'org.slf4j:slf4j-jdk14'
@@ -72,7 +72,7 @@ dependencies {
     baloTestImplementation project(path: ':ballerina-lang:value', configuration: 'baloImplementation')
     baloTestImplementation project(path: ':ballerina-lang:xml', configuration: 'baloImplementation')
     baloTestImplementation project(path: ':ballerina-lang:bool', configuration: 'baloImplementation')
-    baloTestImplementation project(path: ':ballerina-lang:compilertest', configuration: 'baloImplementation')
+    baloTestImplementation project(path: ':ballerina-lang-test', configuration: 'baloImplementation')
     baloTestImplementation project(path: ':ballerina-lang:internal', configuration: 'baloImplementation')
 }
