+++ settings.gradle
@@ -17,7 +17,7 @@ include(':ballerina-lang:typedesc')
 include(':ballerina-lang:value')
 include(':ballerina-lang:xml')
 include(':ballerina-lang:bool')
-include(':ballerina-lang:compilertest')
+include(':ballerina-lang-test')
 include(':ballerina-langlib:test')
 include(':stdlib')
 include(':ballerina-auth')
@@ -147,7 +147,7 @@ project(':ballerina-lang:typedesc').projectDir = file('langlib/lang.typedesc')
 project(':ballerina-lang:value').projectDir = file('langlib/lang.value')
 project(':ballerina-lang:xml').projectDir = file('langlib/lang.xml')
 project(':ballerina-lang:bool').projectDir = file('langlib/lang.boolean')
-project(':ballerina-lang:compilertest').projectDir = file('langlib/lang.compilertest')
+project(':ballerina-lang-test').projectDir = file('langlib/lang.test')
 project(':ballerina-langlib:test').projectDir = file('langlib/langlib-test')
 project(':ballerina-runtime-api').projectDir = file('stdlib/runtime-api')
 project(':ballerina-runtime').projectDir = file('bvm/ballerina-runtime')
